abstract class, delete on edit doest work yet
